Save memory

+++ src/+burst/prep1.m
@@ -23,7 +23,7 @@ function [dat,opts] = prep1(p0,f0,rgT,opts,ff)
     if exist('rgT','var') && ~isempty(rgT)
         dat = dat(:,:,rgT);
     end
-    dat = double(dat);
+    dat = single(dat);
     maxDat = max(dat(:));
     dat = dat/maxDat;
     dat = dat(bdCrop+1:end-bdCrop,bdCrop+1:end-bdCrop,:);
